Update application manifest and refactor source code for improved stability and performance
*   Updated `AndroidManifest.xml` to include fullscreen theme, singleTask launch mode, and portrait orientation for the main activity.
*   Replaced default launcher icons with `pticon`.
*   Removed unused `activity_main.xml` and various mipmap/drawable assets.
*   Refactored `Papertoss.java`, `Menu.java`, and `Level.java` to improve transition logic, state handling, and floating-point calculations.
*   Optimized `Evt.java` by replacing `Iterator` with an enhanced for-loop for listener notification.
*   Simplified exception handling in `Scores.java`, `SaveData.java`, and `SoundMgr.java` by removing redundant `throws` declarations and try-catch blocks.
*   Converted a lambda expression to an anonymous inner class in `PapertossGLSurfaceView.java`.
*   Added `JADX INFO` comments across various source files.

+++ app/src/main/java/com/bfs/papertoss/platform/Random.java
@@ -1,6 +1,6 @@
 package com.bfs.papertoss.platform;
 
-/* loaded from: classes.dex */
+/* JADX INFO: loaded from: classes.dex */
 public class Random {
     public static int randomi(int low, int high) {
         int randi = (int) (Math.random() * 2.147483647E9d);
@@ -8,7 +8,7 @@ public class Random {
     }
 
     public static float randomf(float low, float high) {
-        return (float) ((Math.random() * (high - low)) + low);
+        return (float) ((Math.random() * ((double) (high - low))) + ((double) low));
     }
 
     public static void main(String[] args) {
